Factor out logoutHandler
